More updates for marking MathObjects with the context in which they were created, so they should now work like the Formula objects in that respect.  As they are combined via overloaded operators, they should pass on their contexts correctly.
Also normalized the use of getFlag() to obtain flags from the
MathObject's context rather than looking in the context directly.
This allows the math object to override the flag by setting the flag
value in the object's hash (e.g., $f->{tolerance} = .001).  I've also
added the ability to override context flags via the answerHash (e.g.,
$f->cmp(tolerance => .001)), though some filtering may need to be
added to this at some point.  Note that ONLY the context flags can be
overridden, not other parts of the context.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@4979 c0722133-6baf-4dd8-8699-98d999cd4f06