This is a temporary hack that allows sticky answers to preserve all the characters the student typed.  PGbasicmacros.pl currently removes several characters, including $, which is needed for monitary answers.  Eventually, PGbasicmacros.pl should be fixed, but for now this will work, and it should not cause trouble after PGbasicmacros.pl is fixed.  In the end, this file can be emptied, then it should issue a warning message, and finally it can be removed.
The hack is to copy the inputs_ref and modify the answers so that the
PGbasicmacros.pl macros will not remove those characters (the key is
to convert them to HTML entities instead).  Then after the problem has
been processed, we put the original answers back so that the answer
checkers will work properly on them.  This is accomplished by
replacing ENDDOCUMENT with a new routine that first puts back the
answers and then calls the old ENDDOCUMENT.  This has the unwanted
side effect of putting error messages into the error log, but I can't
seem to work around that.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@5342 c0722133-6baf-4dd8-8699-98d999cd4f06