This is a rewrite of the Inequalities context that makes a number of improvements over the older version.
First, the error messages are much better when students enter
incorrect answers.

Second, inequalities and intervals are now separate classes and so you
can enter both types of notation and they will stringify and texify
correctly, so there is no need for the stringifyAsInequalities flag
any longer.

Finally, there is a new Inequality() constructor that can be used to
coerce Intervals, Sets, and Unions to become their corresponding
inequalities, and vice versa, Interval(), Set() and Union() can be
used to coerce an Inequality to be represented in its interval-style
notation.

This is not backaward compatible with the ealier version, but since
that is only a couple of weeks old, I don't think it should be an
issue, as very few problems will have been written for it in that
time.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@5434 c0722133-6baf-4dd8-8699-98d999cd4f06