Updated contexts to include a "name" field that at least tracks what context you started with (though it can be modified and no longer be the same as the original context).
Remove the individual named variables in the Default.pm file; they are
now available only through the %Parser::Context::Default::context
hash.

Remove the >< and . operators, the <...> parentheses, the norm and
unit functions, and the i, j, and k constants from the Point context.
So the Point context no longer includes vectors and vector
operaterations.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@5439 c0722133-6baf-4dd8-8699-98d999cd4f06