Changed plot_functions so that it uses the Formula MathObject to parse the the function in
$f1 = qq! x^2 - 3*x + 45 for x in [0, 45) using color:red and weight:2!

instead of using the built in perl parser.

This allows things such as $formula=Formula("|x|");

$f1 = qq! $formula  for x in [0, 45) using color:red and weight:2!

to be evaluated. There should be no apparent change to old problems
since the Formula MathObject parser handles a superset of the builtin perl parser syntax.

This is not yet the most efficient method to have GraphObjects and
Formula's interact since the same string is being recompiled several times,
but it is a useful hack which preserves backward compatibility.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@5569 c0722133-6baf-4dd8-8699-98d999cd4f06