Inequalities that are evaluated and turned into intervals now properly retain the order of operands (x > 10 remains x > 10 rather than becoming 10 < x).
Unions of the form "x < a or x > a" will be displayed as "x != a" instead,
and "a < x < b or b < x < c" will be shown as "a < x < c and x != b"
instead.  This behavior is controlled by the showNotEquals context
flag.  Setting it to zero prevents conversion of unions to their
not-equal forms.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@5951 c0722133-6baf-4dd8-8699-98d999cd4f06