This version of the tooltip fix is still temporary but it works slightly better in practice than the fix just committed.
This method first attempts to calculate the latex string from {correct_ans} and if that fails
it attempts the mathObject->Tex, and finally it places an empty string in {correct_ans_latex_string} if 
all else fails.  With this method one gets the typeset version of pi  rather 
than the numerical result for the correct answer for Compute("pi").



git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@6319 c0722133-6baf-4dd8-8699-98d999cd4f06