the AddToTeXPreamble()  command now adds short commands (such as \newcommand{\foo}{bar} )
Lines such as 
AddToTeXPreamble("\\newcommand{\\myVec}[1]{ \\overrightarrow{#1}  }" );

are best  added in PGcourse.pl so that the style of \myVec for writing vectors
is determined for the whole course.

You could also use it in a single problem, but the style would not 
be made for hardcopy. If it were and you used the statement in two problems
there would be a "redefine error".




git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@6863 c0722133-6baf-4dd8-8699-98d999cd4f06