Added one more hack so that both warning messages and debug messages are moved from PGcore and PGalias back to Problem.pm and printed to the screen.
A better fix will be to create a storarge array in PG.pl and pass that address
to PGcore when it is created for use in storing warning (resp. debug) messages.
When PGalias, and the other PG subclasses of PGcore are created they will also be pointed
to the same storage location by their parent.