Move queue for warning/debug messages to top level of PGcore (out of ->{flags})
Earlier the warning/debug queues for PGalias were tied to those for PGcore so
no transfer of messages is needed. This still needs testing.

FIXME: Translator.pm, in the commented out statements, should be using the methods
of PGcore to access messages rather than doing it directly.  This seems to be
ok in PG.pl and also in Problem.pm where these messages are retrieved.