Fix bug in pretty_print that allowed large recursion
The levels variable was not being passed.  Also upped the level to 5
and prevented from any of the contents of a PGalias object being
displayed -- it takes over.  (Preventing just the envir object
from being displayed might be enough. )  Also the effect is worse
in TeX then in HTML mode.  Perhaps we could let PGalias be displayed
to a certain depth.

PGalias might have an infinite self reference.  When the level
variable was not being passed the entire server would hang during
pretty_print.