Fix includePGtext to include the current preprocessing routine defined for the Translator.
The earlier version had copy/paste code that preprocessed the BEGIN_TEXT/END_TEXT string but
none of the other strings added to the default_preprocessing_code subroutine.  The current
version actually directs includePGtext() to use the preprocesing code defined in
$translator->{preprocessing_code}.  Hence includePGtext will be consistent with the
translator of the current problem.

Preprocessing must be repeated in includePGtext because the string that is added by
includePGtext has not been run through the preprocessing subroutine.

includePGtext is used by includePGProblem and by includePGfile.
It seems possible that the union macro includePGfile can be folded into includePGProblem
which is defined in PG.pl