Use &$PREPROCESS_CODE to process strings in includePGtext
This was originally added in July 15, 2012 but was subsequently removed.
This also removes references to includePGtext in Translator.pm where it does not belong.
The only location for includePGtext is in  PG/IO.pm.  It is referred to from PGcore.
The reference to includePGtext in PG.pl for the includePGproblem macro now references the PGcore version.