Modify knowlLink to accommodate updated version of knowl.js
Updated knowl.js from the AIM website -- allows HERE doc version of knowls.

I was going to implement this change myself but this version from AIM
beat me to it, does it better -- and it's the standard distributed version.

There are accompanying changes to knowlLink in PGbasicmacros.pl to
accomodate the new version.