Added a mode for pretty_printing to a standard text file.
This is not yet used for anything in PG.  I thought it would be useful for printing out the trace of the answer evaluators, but
using DEBUG_MESSAGE and the html mode turned out to work better.

This pretty_print_text() routine is essentially the same as several of the pretty_print_rh routines that are used in clients/renderProblem.pl
WeBWorKClients.pm and other files on the WeBWorK2 side.  All the uses of pretty_print on the PG side seem to be going through PGUtils.pm