Create contextTrigDegrees.pl
A new macro that defines a context TrigDegrees where trig functions are defined in degrees.  Using this context, trig functions will be evaluated in Perl mode as well as within MathObjects.  Derivative methods are defined on these trig functions as well.  This version of contextTrigDegrees.pl does not have the capability for using these trig functions in vector or matrices.