Use PGanswergroup to match answer evaluators with their attached answer blanks.
This works for many answer evaluators -- perhaps not yet all.  More work needs to be done to clean up the code in this file.  After that it will be possible to change the way that some of the answer blanks (matrices and multi answer) are created and labelled so as to make the process of evaluating the student answers saner.