Hacks to fix the failure to accept ',   $  and _   (as well as hacks to fix the previous problems).
There should be a more permanent fix but I'll have to think about it.  The original translator worked because it went back
and looked at the original values in $inputs_ref  before any filtering took place.  The current fix of selectively undoing some
of the HTML entity translations works for now.