add 'dir="auto"' to <input> elements
This assures that the direction is automatically detected. In a global
RTL setting entering ASCII characters should lead to a LTR direction. If
entering characters in languages that are traditionally written in the
RTL direction, the direction should be RTL.

See also: https://www.w3schools.com/tags/att_global_dir.asp

An alternative setting would be "ltr" for the moment, since most input
will be left to right. But this would probably lead to problem in case
of problems that require the student to write down a proof or some other
form of "free text" in languages that are written from right to left. So
I propose the "auto" setting.