To improve the safety of adding the the balancing sanity check feature to the main code-base, conditional code was added to totally bypass the new code when so desired.
Setting the key
	warn_on_internalBalancing_errors
in the flags hash defined in lib/PGcore.pm to 0 is the recommended
way to trigger the bypass.

This hash value is used to set the variable
	$warn_on_internalBalancing_errors
in macros/PG.pl, which is what actually is checked in the conditional
code (the if conditions).

In particular, with the bypass setting, the block of code in EV3P()
(inside macros/PGbasicmacros.pl) where regular expression code is used
to modify the original code of the PG file in order to add in calls
to the tracking functions is skipped.