wantarray issue
The issue is explained in the thread for d61e0c7. Before the change here, `$ans->$method(@options)` is the argument to `PGML::LaTeX()`, therefore an array. Within certain methods, wantarray is present, and detects that `$ans->$method(@options)` is supposed to be an array, and behaves differently. This edit makes `$ans->$method(@options)` used in scalar context, changing the reaction from wantarray. And this reverts the behavior to how it was before d61e0c7.