Add a missing call to INSERT_RESPONSE to the PopUp MENU method in the case that it is an extended answer.
Currently, if a PopUp answer is used as a later answer in a MultiAnswer
problem with the singleResult option, then the answer for that part is
not saved.  When answers are submitted the selected answer initially
displays when the page reloads, but if you navigate away from the
problem and come back, the answer to that part is no longer there.  The
reason for this is that the answer name is never appended to the
response group as is done in the NAMED_ANS_RULE_EXTENSION method.