Add macros to allow insert start/end DIV and SPAN tags in HTML mode.
The current primary use of this is to support changing between RTL
and LTR text directions using the HTML dir attribute of SPAN for
inline changes and DIV for block level changes, as needed.

The commands can accept optional strings which will be inserted
in LaTeX and preTeXt mode, to allow passing suitable commands to those
output frameworks also, as is expected to be needed. The use of the
optional strings has not yet been fully tested.

Support for Hebrew, Arabic, Persian, and other right-to-left languages
requires changing to LTR mode for parts of the question text, such as
blocks of math mode and answer input elements. The existing BLTR and ELTR
tags are a far more restricted method of getting such an effect, and
were not designed to provide hooks for sending LaTeX code to be
used for the LaTeX output mode.

The code also allows setting the HTML lang attribute for when a language
change occurs (required by WCAG 2.0).

There is also support for setting the class and style attributes, which
would allow problem authors to use these attributes to control the styling
of HTML output. Use to control the class and style attributes has not yet
been tested much.