Fix for a bug when seen on the first load of a problem with a radio button used in MultiAnswer. It was reporting:
substr outside of string at line 375 of [PG]/macros/parserRadioButtons.pl
Use of uninitialized value $index in array element at line 376 of [PG]/macros/parserRadioButtons.pl
Use of uninitialized value $index in array element at line 377 of [PG]/macros/parserRadioButtons.pl

The root problem is that labelText() is apparently sometimes called
with either an empty second ("index") argument or one too short to
allow the code
	$index = substr(shift,1);
to work. The fix defers the call to substr until it is certain that
it will work, and when it would fail - returns a fixed string "Bn".