Revised fix for a bug when seen on the first load of a problem with a radio button used in MultiAnswer when allowBlankAnswers is set to 1.
In such cases, on problem load, and when no radio button was selected
on an answer submission the following warning message was issued:

substr outside of string at line 375 of [PG]/macros/parserRadioButtons.pl
Use of uninitialized value $index in array element at line 376 of [PG]/macros/parserRadioButtons.pl
Use of uninitialized value $index in array element at line 377 of [PG]/macros/parserRadioButtons.pl

The root problem is that when no buttom has been selected,
labelText() is being called with a second ("index") argument which
cannot be handled by the code
	$index = substr(shift,1);
which expects a string of the format Bn where n in an integer index.

An initial fix I proposed as https://github.com/openwebwork/pg/pull/405
was rejected as being a bad solution to the problem.

The revised fix was proposed by Davide P. Cervone and avoids the call
to labelText() when no answer was selected.