Add context flag multiplyRightToLeft
Add a context flag `multiplyRightToLeft` with default value `0` that does not change existing behavior, and also document this feature.  Many textbooks use the convention that permutations are functions that are composed from right-to-left, which is why this feature is useful for problem authors.  Adding this flag should not affect any of the existing webwork problems that use this macro because the default behavior (multiply left-to-right) is the same as before.  For instance, we have `(1 2 3)(1 2) = (2 3)` with the default setting `multiplyRightToLeft=>0`, whereas we have `(1 2 3)(1 2) = (1 3)` when `multiplyRightToLeft=>1`.