Create GroupTableAnswer.pl macro
Allows multiplication (group tables) to be created and graded with ease.  An editable html table is used to collect answers (instead of html form inputs) and javascript is used to get the student answers from the table and format them as a MathObject matrix on submit, and to update the table entries using the previous answers on load.