Display answer as a table instead of a list
Use a post-filter to modify the answer hash so that it displays a table to the student, rather than a list, for the entered answer, answer preview, and correct answer.  Note: this works properly except when a group table is used inside of a scaffold section (see bug #4430 about scaffold.pl in WeBWorK's Bugzilla) because for some reason scaffold is processing the post-filter twice.