Fix the sage.pl macro.
First, there was an obvious typo that made the macro's ShowAnswerBlank
option fail.

Next, sage uses python3 which requires that the print function have its
arguments parenthesized.  This causes the record_answer method in the
ShowAnswerBlank eq 'none' case to fail.

Next, '<script src="$CellServer/static/jquery.min.js"></script>' was
recently added to this macro before the loading of the sagecell
javascript.  The sagecell javascript internally loads its own jQuery
instance, and so there is no need for that.  Furthermore, that version
of jQuery conflicts with the version currently in use by WeBWorK.

Finally, the sagecell server does not correctly deal with jQuery.  They
recently added the usage of the noConflict method, but then they still
use and modify the global $ variable.  This conflicts with our jQuery
usage.  So our version of jQuery is cached and restored after the
sagecell javascript is loaded and executed.

See https://webwork.maa.org/moodle/mod/forum/discuss.php?d=5011 for some
of my discussion on this.