The stringify_answers method of Translator.pm now JSON encodes the mathQuillOpts key of the answer hash if defined and a hash reference. This allows the value of that key to be a hash.  This is easier to work with than manually creating JSON strings.  Furthermore, it can be directly passed to the mqedit.js javascript as a data attribute with an object value.
So this changes the way the mathQuillOpts answer hash key works.
Previously to disable the spaceBehavesLikeTabs option (i.e. to enable
typing of spaces) and the disable rootsAreExponents you would need to do
  `$ans->cmp(mathQuillOpts => "spaceBehavesLikeTab: false, rootsAreExponents: false")`
Now you will do
  `$ans->cmp(mathQuillOpts => { spaceBehavesLikeTab => 0, rootsAreExponents => 0 })`

Note that you can also use `$ans->cmp(mathQuillOpts => "disables")` to
disable MathQuill answers for an answer blank.  (This is done in the
webwork2 code).

This also automatically disables MathQuill answer boxes when the answer
is a parserWordCompletion.pl answer.