Implement a better method for generating svg images with TikZImage.pm.
The current implementation generates a pdf using pdflatex, and then
converts the pdf to svg using imagemagick's convert utility.

This changes it so that it uses latex to generate a dvi with the
pgfsysdriver redefined to pgfsys-dvisvgm.def, and then converts the dvi
to svg using dvisvgm.  This is the best way that I have found of
creating svg images from tikz/tex.

This does require another external program (dvisvgm).  That is added in
a corresponding pull request to webwork2.