Attempt to fix MathQuill issues with mixed numbers, units, and scientific notation.
For this switch to inserting mathquill responses on the PG side of
things, and add a special precedence that is used only for student
answers and only when MathQuill is enabled for the answer.

The mathquill latex answers are now stored in KEPT_EXTRA_ANSWERS instead
of as an additional response appended to the answer.

This requires changes to MathQuill and for webwork2 (or the front end)
to pass in an environment variable via the translation options that
tells PG that MathQuill is enabled.  These changes are in a
corresponding pull request to webwork2.

Special handling of MultiAnswer's is implemented.  It is now possible to
disable a single part of a multi answer, instead of only having the
option to disable all parts.

One of the upsides of adding mathquill responses here is that we have
access to an answer value's context.  One use is implemented here.  That
is to make MathQuill automatically disable the rootsAreExponents option
if the root function is available in the context for an answer.