Set the defaults in pg_defaults.yml.dist to be the same as they are in site.conf.dits, and add the external program tar which is used by TikZImage.pm.
Delete the commented out code in lib/WeBWorK/PG/IO.pm that was replaced
with new code in this pull request, and the commented out obsolete code
in lib/PGcore.pm.

Run perltidy on PGEnvironment.pm using the same settings we use for ww3.
That is a new file, so it should start clean.  Also update the license
in that file.

Revert the changes to lib/chromatic/color.c.  Why was that file changed?