Add the files from webwork2 htdocs that are used by pg problems to a new htdocs directory here.  All of those files have been deleted from the webwork2 repository by a corresponding pull request.
Note that a package.json file has been added that installs jsxgraph.
This is not needed for webwork2 as jsxgraph has also been added the the
package.json file there.  So a webwork2 installation step is not added.
This will be needed for the standalone renderer and webwork3 though.
The parserGraphTool.pl macro has been updated for this change of
location.

Also the knowl.js and knowlstyle.css files are moved to
htdocs/js/apps/Knowl in the pg htdocs directory.