Change the way that coloring of inputs is done for correct and incorrect answers.  Instead of creating a global javascript variable, use data attributes.
Also, instead of using inline javascript to make clicking on the
correct/incorrect link set focus on the input, add an event listener.
This is done from htdocs/js/apps/InputColor/color.js, so perhaps that
file is not properly named anymore.

Note that if MathQuill is enabled this is handled for mathquill inputs
in htdocs/js/apps/MathQuill/mqeditor.js.  It did so before, but in a
hackish way.  Now it is much cleaner.

As with all of these changes this is also designed to work with webwork3
and the standalone renderer.