Add stub methods for FlashApplets, JavaApplets, and CanvasApplets that show a warning if these methods are called.  So if a problem in the OPL is used that uses one of these deprecated applet types, intructors will know not to use the problem.
Also stub out the findAppletCodebase method in PGalias.pm.
None of those "codebases" will work anymore, and soon they will be
deleted (via openwebwork/webwork2#1553) which will cause this method to
fail with errors.  Interestingly some of the Geogebra problems in the
opl that call this method still work, even with this stub.