Add unit tests (#618)
This sets up some unit tests for some PG macros and contexts. It also provides a framework and examples for how to do further unit testings.

Also, there is a GitHub action to run the tests on pushes to GitHub.  Note that unit tests are disabled for the repository until this workflow action can be updated.  Currently it is failing.

Co-authored-by: Peter Staab <peter.staab@gmail.com>
Co-authored-by: Fabian Gabel <fabian.gabel@tuhh.de>
Co-authored-by: Michael Gage <1203580+mgage@users.noreply.github.com>