Fix another regression caused by #609.
If a problem sets the answer to be a formula with an assigment,
(something like `Formula("x = 5")` then all answers are counted
incorrect and a warning is displayed that the evaluated answer is not an
answer hash.  This worked prior to #609.  The cause of this was the
removal of a line that should not have been removed that converts the
other answer being compared to to a Formula if it is not already.

Note that the original issue that was attempted to be fixed by #609
still is fixed with this change.

This fixes issue #644 that I just submitted about this.