Add an asset build system for javascript and css files.
All javascript and scss/css files in htdocs/themes and htdocs/js/apps are compiled to minimized js or cs files.  The filename of the generated file contains a hash of the contents.  The generated file names are saved to the file `htdocs/static-assets.json`, each identified by the unminimized file name. This is all done by the `generate-assets.js` script that is executed when `npm install` is run in the `htdocs` directory.  See the corresponding webwork2 pull request for more details of how this works.

The parserGraphTool.pl no longer requests the `graphtool.min.js` file which has been deleted.  Of course, now that file (with content hash included in the file name) will be used via the lookup.