Rewrite the graphtool javascript using modern javascript and remove the one last usage of jquery.
Also run perltidy on the macros/parserGraphTool.pl file, and do a little
reorganization to keep the point tool code together.

Add new options `xAxisLabel` and `yAxisLabel` that can be set to change
the labels that are displayed on the horizontal and vertical axes.
These can also be set to the empty string to remove the labels.

Remove the `fixed` attribute from the highlight points of the tools.
JSXGraph does not snap to grid if this option is set.

For the point graph object its point needs to also be a defining point
of the object.  This makes it so that a point can not steal focus from
another object if it shares the same location as a defining point of
that object.  Sorry Alex I think I guided you incorrectly here.