Fix tab order for the three point quadratic and four point cubic tools.
There was an indexing error in the logic of the tab focus handler of the selection tool that results in the last defining point of an object that has three or more defining points not working correctly.

To see the issue, graph any object and then graph a quadratic or cubic after that.  Immediately after completing the graph of the quadratic or cubic hit shift-tab.  With the code in the release candidate branch the focus will jump back to the last point on first object that was graphed.  With this pull request the focus will go to the next to last point on the quadratic or cubic as it should.