Fix pointer focus for the center of a circle.
Curently the center point of a circle can not be focused by clicking on
it with the mouse.  This is because the jsxgraph circle hasPoint method
only returns true if the circle itself contains the given point, but not
if the center point contains the given point.  So redefine the hasPoint
method for a circle to fix that.