Wrap solutions and hints in <div>...</div>.
Don't just start a <p> tag, and never end it, and it seems that a <p>
tag is the wrong thing in any case.

This doesn't fix the bigger issue of the $PAR usage.  Problems in the
OPL use this frequently.  The problem is that $PAR starts a <p> tag, but
never ends it.  This results in invalid html.