Move the WeBWorK::PG module to PG and improve PG's local configuration.
Remove the ALWAYS_SHOW_HINT_PERMISSION_LEVEL and
ALWAYS_SHOW_SOLUTION_PERMISSION_LEVEL settings.  Whether hints and
solutions are shown is now determined by the front end.

The pg flag/variable $showHint has been removed.  The front end has
complete control on when hints are shown.  Not the problem author.

Also the options to use knowls for solutions and hints have been
removed.  Always use knowls for solutions and hints.

Showing the various pieces of debugging information for a problem is now
controlled with translation options.  The front end is responsible for
converting permissions into these debugging options.  The
VIEW_PROBLEM_DEBUGGING_INFO special pg environment variable was removed
and is one of these debugging options.

Many unnecessary PG environment variables have been removed.  One of
these is the session key.

Remove reduced scoring entirely from PG.  That is the responsibility of
the front end to implement.

Pg now has not dates and and no permissions.

The color.c file has been removed.  The method in that file has been
reimplemented in Chromatic.pm using perl.  It is perhaps a bit slower,
but not significantly in the scope of rendering a problem.  This can be
tested with the problems
Library/NAU/setGraphTheory/FindChromaticNumber.pg,
Library/NAU/setGraphTheory/FindRoomColor.pg, and
Library/NAU/setGraphTheory/RoomSchedule.pg
which happen to be all known problems that use the Chromatic module.

The image generator has been rewritten to remove the dvipng_align hack
and to allow the usage of sqlite for a database backend for storing
dvipng depths.

Note that PGalias no longer uses the course name, user login, set
number, and problem number for generating resource UUID's.  It now only
uses the problem seed, psvn, and problemUUID.  It is the caller's
responsibility of ensuring that problemUUID guarantees the necessary
uniqueness.