Add some unit tests.
The t/macros/load_macros.t file tests that all macros load without
errors or warnings.

There is one macro that will not work in this test and it is skipped.
That is the answerDiscussion.pl macro.  It requires a file that is not
part of the PG distribution.  It also has invalid `defined(@array)`
calls.

Warnings found in other macro files were fixed.

The PG_module_list.pl macro was deleted.  That is clearly obsolete and
will not work.

There are also some tests added in t/pg_problems that test that
rendering with the WeBWorK::PG module works and demonstrate how to use
it in unit tests.