Fix graphing of the third point for the GraphTool cubictool.
Currently if you graph three consecutive points where each point is one
unit to the left of the previously graphed point, then when the third
point is graphed there will be an error that causes the graph tool to
become mostly unresponsive.  The problem was an undefined variable being
used in phase3 of that tool.