Disable redefine warnings when macro files are loaded.
Many subroutines are redefined in macros.

In addition if multiple WeBWorK::PG::Translator instances are created in
the same process, the safe compartment is not entirely isolated between
instances, and this causes redefine warnings if a macro is loaded in
both instances.  This is due to a memory leak somewhere and needs to be
fixed.