Restore the functionality of the LiveGraphics3D.pl macro.
It turns out @taniwallach was correct in that the javascript conversion
method does work.  See the discussion in #728.

There is still some more work that is needed here (cleanup really).  The
se_javascript_for_live3d option needs to be removed, and that needs to
be always used.  Setting that option to 0 will fail in any case.  This
pull request makes these problems work at least.