Add a script to perltidy PG problem files
The script just runs perltidy but it preprocesses the code in the file
with the default_preprocess_code of the translator.  After that the code
is valid perl that perltidy works on.  Then it postprocesses the file to
change the code back to PG (with BEGIN_TEXT/END_TEXT,
BEGIN_PGML/END_PGML, etc).

The script accepts all arguments that perltidy does.  Note that by
default the script attempts to use the .perltidyrc file in the PG root
directory.  If the PG_ROOT environment variable is not set or the script
can't find the PG root directory for some reason, then perltidy defaults
are used.  You can also explicitly use a different profile by passing the
perltidy -pro argument.